Refactor curl and downloader code.
This is NFC refactoring. I have tried to make the code clearer and use
type system to provide some guarantees. Specifically, when the easy
handle is used by by multi handle, it is converted into curl_easy_in_multi,
which makes more clear that we should not modify the easy in that state.

This also removes quite ugly search for the handle that just finished
downloading. It uses a "private" data in the easy for that, and stored
"Transfer" objects in a std::set instead of a vector.