Do not load images from disk if we have no time left
This makes zooming smoother. Before, when we transition from one zoom
level to another, we would load a lot of images in the same frame, which
would make this frame take too much time. This showed up as jittery
zooming. Now we don't load images when we have run out of time budget
for a frame (currently hardcoded at 12 ms). This makes zooming smoother
but makes map flicker, because some images will be missing for a few
frames. To solve this, we will need to use images from other levels
as a fallback instead of missing images.